layered-sets can now be made via texref
Next I want to add things like pull-g for layered-sets. Right now the
api feels a bit spare.

It would be nice to work out how to represent a slice of a 3d texture
so I could pull-g just that. The pulling is easy, it's the CEPL
representation that is eh. That is due to not being able to subseq-g a
texture and thus not being able to refer to that region.